benchmarks: split benchmarks for avoid depends unnecessary packages
