+++ go.mod
@@ -3,8 +3,6 @@ module github.com/minio/simdjson-go
 go 1.15
 
 require (
-	github.com/buger/jsonparser v1.1.1
-	github.com/json-iterator/go v1.1.9
 	github.com/klauspost/compress v1.13.6
 	github.com/klauspost/cpuid/v2 v2.0.9
 )
